remove characters
