+++ src/components/App.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import Header from '../containers/Header';
-import Lists from '../containers/Lists';
+import EpisodesList from '../containers/EpisodesList';
 
 const App = () => (
   <div className="App">
     <Header />
-    <Lists />
+    <EpisodesList />
   </div>
 );
