--- src/constants/InitialState.js
@@ -1,3 +1,0 @@
-const VIEW = 'episodes';
-
-export default VIEW;
