+++ src/containers/EpisodesList.js
@@ -1,0 +1,45 @@
+import React, { useEffect } from 'react';
+import PropTypes from 'prop-types';
+import { connect } from 'react-redux';
+import Episodes from '../components/Episodes';
+import { getEpisodes } from '../actions/index';
+
+const EpisodesList = ({
+  episodes, getEpisodes,
+}) => {
+
+  useEffect(() => {
+    getEpisodes();
+  }, [getEpisodes]);
+
+  return episodes.length === 0 ? <div>Please wait</div> : (
+    <div>
+      {episodes.map(episode => (
+        <Episodes 
+        key={episode.id} 
+        name={episode.name}
+        season={episode.season}
+        number={episode.number}
+        airdate={episode.airdate}
+        runtime={episode.runtime}
+        image={episode.image.medium}
+        summary={episode.summary} />
+      ))}
+    </div>
+  );
+};
+
+const mapStateToProps = state => ({
+  episodes: state.episodes,
+});
+
+const mapDispatchToProps = dispatch => ({
+  getEpisodes: () => dispatch(getEpisodes()),
+});
+
+EpisodesList.propTypes = {
+  episodes: PropTypes.instanceOf(Array).isRequired,
+  getEpisodes: PropTypes.func.isRequired,
+};
+
+export default connect(mapStateToProps, mapDispatchToProps)(EpisodesList);
