+++ src/reducers/index.js
@@ -1,8 +1,6 @@
 import { combineReducers } from 'redux';
 import episodes from './episodes';
-import cast from './cast';
 
 export default combineReducers({
   episodes,
-  cast,
 });
